feat(config): move help and validation to platform
- Move help_section_for_flag(), help_text(), print_help(), and validate_args() from NES config to platform config
- Move all_cli_flags() to platform to support help and validation
- Make NES CLI_FLAGS pub(crate) so platform can access it
- Make NES config module pub(crate) in console mod
- Update all references (main.rs, tests) to use platform versions
- Remove unused std::fmt::Write import from NES config
- All 8666 tests pass
- --help output is byte-for-byte identical
- Clippy clean with -D warnings
- WASM tests pass (54 tests)

Completes sub-issue #2159
